chore: use Mocha
